+++ src/main/java/net/spell_engine/client/gui/SpellTooltip.java
@@ -431,6 +431,12 @@ public class SpellTooltip {
                             }
                         }
                     }
+                    case TAUNT -> {
+                        if (impact.action.taunt != null) {
+                            var taunt = impact.action.taunt;
+                            addToken("taunt_chance", percent(taunt.chance), tokenReplacements);
+                        }
+                    }
                 }
             }
             var area_impact = spell.area_impact;